requirements.txt
altair==5.4.1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
boto3==1.35.55
botocore==1.35.55
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
comm==0.2.2
debugpy==1.8.7
decorator==5.1.1
executing==2.1.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.13.2
nest-asyncio==1.6.0
numpy==2.1.3
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.19.0
prompt_toolkit==3.0.48
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
s3transfer==0.10.3
six==1.16.0
smmap==5.0.1
stack-data==0.6.3
streamlit==1.40.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vaderSentiment==3.3.2
wcwidth==0.2.13
